server/requirements.txt
alembic==0.9.2
appdirs==1.4.3
Babel==2.4.0
blinker==1.4
click==6.7
decorator==4.0.11
Flask==0.12.1
Flask-Babel==0.11.1
flask-cors
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemy==0.56
geopy==1.10.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pbr==2.0.0
pyjwt==1.4.2
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.9
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Werkzeug==0.12.1
Whoosh==2.7.4